backendApp/requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0